#75 Add possibility to join for new validators. Move hooks in separate file.
